+++ Cargo.toml
@@ -1,6 +1,9 @@
 [workspace]
 members = [
   "./other/light-state-domain-driven",
+  "./other/rust-fundamentals-engineered-from-pl-uni/z1a1",
+  "./other/rust-fundamentals-engineered-from-pl-uni/z1a2",
+  "./other/rust-fundamentals-engineered-from-pl-uni/z1a3-4",
   "./pasja_informatyki/odc-4/lotto_game",
   "./pasja_informatyki/odc-5/evaluation_mean",
   "./pasja_informatyki/odc-5/fibbonaci_counter",
@@ -22,5 +25,5 @@ members = [
   "./small_test_tools/restart_swap",
   "./small_test_tools/test_curl_vs_reqwest",
   "./small_test_tools/texty-rusty-sound-"
-, "other/rust-fundamentals-engineered-from-pl-uni/z1a1", "other/rust-fundamentals-engineered-from-pl-uni/z1a2", "other/rust-fundamentals-engineered-from-pl-uni/z1a3"]
+]
 resolver = "2"
