+++ other/rust-fundamentals-engineered-from-pl-uni/z1a3-4/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "z1a3"
+name = "z1a3-4"
 version = "0.1.0"
 edition = "2024"
